// src/api/orval/model/userDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface UserDto {
  id?: number;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  email?: string | null;
  isAdmin?: boolean;
  isBlocked?: boolean;
  hasDashboardAccess?: boolean;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  providerRowId?: number | null;
  createDate?: string;
  lastUpdate?: string;
}